Filter버튼 리스트 컴포넌트 재구성 - SearchScreen에 전체 UI 구상하지 않고 컴포넌트 나누어서 재구성 - SearchScreen - Components - SelectFilter - Components(ChipList,Data)
