+++ Screen/Main/Screen/Search/SearchScreenStyle.tsx
@@ -3,7 +3,6 @@ import {StyleSheet, ViewStyle} from 'react-native';
 type Style = {
   container: ViewStyle;
   searchBar: ViewStyle;
-  chipView: ViewStyle;
   filterView: ViewStyle;
   selectedChipView: ViewStyle;
 };
@@ -23,10 +22,6 @@ const styles = StyleSheet.create<Style>({
     top: '3%',
   },
   filterView: {top: '10%'},
-  chipView: {
-    margin: '5%',
-    flexWrap: 'wrap',
-  },
 });
 
 export default styles;
